Connection and Organization Collection Creation
Initial commit with proper connection to our MongoDB Atlas (Cloud) server.

Our connection URL is commented next to the function in which it is used. Currently, the .env file is in the gitignore portion of our backend code.

Once connection was established, the organizationData collection was created in the models.js file.