Created Organization API with GET
Created the API get function to retrieve the all the data in the organization table. Also created an API for getting the organization by the ID. The models.js has a new route added to it, which is the route to the organization.js file. I also changed a variable name in the organization collection.