Graph & Table Data API
Previously we had 2 separate GET API calls for getting number of attendees for an event & for retrieving events in the past 2 months. I created a new API call to essentially merge these two. The new API will only pull information of events in the past two months based on the organization's instance of the application. The results are received in arrays to better help with the graph in the frontend.