Creation of second environment for org 2
Created the second environment for the next organization. The backend changes between the two environments are the organization id that is attached to our cloud and the port number changing from 3001 to 3002 so both instances are able to run simultaneously.  In the front end, the same port number variable was changed and organization id. The static html was changed for the webpage to display the second organizations name. And finally changing the port number for the Vite variable as well to run on port 3002.